topics: Add topic list page
For the topic-list page app bar, we leave out the icon "chevron_down.svg"
since it's related to a new design (#1039) we haven't implemented yet.
This also why "TOPICS" is not aligned to the middle part of the app bar
on the message-list page.

We also leave out the new topic button and topic filtering, which are
out-of-scope for #1158.

The topic-list implementation is quite similar to parts of inbox page
and message-list page.  Therefore, we structure the code to make
it easy to maintain in the future. Especially, this helps us
(a) when we're changing one, apply the same change to the other,
where appropriate, and (b) later reconcile the differences they
do have and then refactor to unify them.

Figma design:
  https://www.figma.com/design/1JTNtYo9memgW7vV6d0ygq/Zulip-Mobile?node-id=6819-35869&m=dev

The "TOPICS" icon on message-list page in a topic narrow is a UX
change from the design.  See CZO discussion:
  https://chat.zulip.org/#narrow/channel/48-mobile/topic/Flutter.20beta.3A.20missing.20topic.20list/near/2177505