theme: Compute colorSwatchFor when subscription is null
This removes the ad-hoc color we use for null subscriptions,
which is a workaround for not having logic to generate different base
colors within constraints for unsubscribed channels:
  https://chat.zulip.org/#narrow/channel/48-mobile/topic/All.20channels.20screen.20-.20.23F832/near/1904009

While this change does not implement that logic either,
it removes the ad-hoc color in favor of a constant base color,
intended for colorSwatchFor.  The base color can be reused on
topic-list page (#1158), giving us access to `.iconOnBarBackground`
and friends for the app bar.