msglist: Set haveNewest from response, like haveOldest
This is NFC with a correctly-behaved server: we set `anchor=newest`,
so the server always sets `found_newest` to true.

Conversely, this will be helpful as we generalize `fetchInitial` to
work with other anchor values; we'll use the `found_newest` value
given by the server, without trying to predict it from the anchor.

The server behavior that makes this effectively NFC isn't quite
explicit in the API docs.  Those say:

  found_newest: boolean

  Whether the server promises that the messages list includes the
  very newest messages matching the narrow (used by clients that
  paginate their requests to decide whether there may be more
  messages to fetch).

  https://zulip.com/api/get-messages#response

But with `anchor=newest`, the response does need to include the very
newest messages in the narrow -- that's the meaning of that `anchor`
value.  So the server is in fact promising the list includes those,
and `found_newest` is therefore required to be true.

(And indeed in practice the server does set `found_newest` to true
when `anchor=newest`; it has specific logic to do so.)