msglist: Always place the sliver boundary at a message
This brings the behavior closer to what we'll want in the future,
and so allows the next commit affecting `middleItem` to be NFC.

This itself is practically NFC.  It's entirely NFC except in the
case where there are no messages, so that `items` consists only of a
history-start marker; and in that case, it causes that marker to go
in the top sliver rather than the bottom.

I believe the only observable effect of that change is that if the
viewport is so very short that the history-start marker is too tall
to fit (so that it's possible to scroll when that's the only item),
then the initial scroll position will be different.