test: Add eg.serverEmojiDataPopular for tests that would break without it
This brings the tests closer to being representative, because this
data (and more) is part of the app's setup for an account on
startup, via fetchServerEmojiData.

For #1495, the app will start depending on this data for the names
of popular emoji. So, prepare for that by filling in the data in
tests that would break without it.