emoji_reaction test: Do a store.setServerEmojiData earlier
The message action sheet is about to condition the appearance of the
reactions row on whether we have ServerEmojiData for any of the
popular emoji. This test-setup code taps "more" on that row to
launch the emoji picker, so the data must be present by the time the
action sheet opens.