msglist: Jump, not scroll, to end when it might be far
When the message list is truly far back in history -- for example,
at first unread in the combined feed or a busy channel, for a user
who has some old unreads going back months and years -- trying to
scroll smoothly to the bottom is hopeless.  The only way to get to
the newest messages in any reasonable amount of time is to jump there.

So, do that.