test: Support testing with poll messages in the message list
When we set up the message list in tests, we do it by preparing the
API response for a message fetch or new-message event, via JSON. But
{Stream,Dm}Message.toJson drops poll data on the floor, which
defeats setting up a poll-style message in the message list. Until
now, anyway, with this workaround.

A reference in a dartdoc to something called
`prepareMessageWithSubmessages` was dangling; it seemed to be
dangling when it first appeared, too, in 5af5c76a8; there's nothing
by that name.