settings: Add language setting
Since there is no Figma design for the settings page yet, the design is
kept simple while matching zulip-mobile: we show both selfname (in its
own locale) and name (in the ambient locale) of each available language
option, and leave out the search funtionality.

We don't allow unsetting the language once it is set, but that can
easily change.

Fixes: #1139