settings: Add language setting
Since there is no Figma design for the settings page yet, the design is
kept simple while mostly matching zulip-mobile: we show both selfname and
name of each available language option, and leave out the search
funtionality.

We don't allow unsetting the language once it is set, but that can
easily change.

Fixes: #1139