app: Override locale with language global setting
This doesn't really have user-facing changes yet, because they cannot
set the language.  In this case, ZulipApp.locale is null, and
localization will follow system setting, like we did before this change.

This behavior is an improvement compared to the legacy app, which just uses
English (en) when language is not set.