message [nfc]: Add _disposed flag; check it
This change should have no user-facing effect.  The one spot where we
have an `if (_disposed)` check in editMessage prevents a state update
anda  rebuild from happening, if the store is disposed before the edit
request fails; the MessageListView with the edited message should get
rebuilt anyway (through onNewStore) when that happens.