l10n: Add zh, zh-Hans-CN, and zh-Hant-TW
Normally, instead of doing it manually, we would add new languages
from Weblate's UI.  In this case, the ones added in this commit are
not offered there.  So we will commit this to GitHub first, and let
Weblate pull the changes from there.

The language identifier zh is left empty, and is set to be ignored
on Weblate for translations.  Translator will be expected translate
strings for zh-Hans-CN and zh-Hant-TW instead.  We can add more
locales with the zh language code if users request them.

See CZO discussion on how we picked these language identifiers:
  https://chat.zulip.org/#narrow/channel/58-translation/topic/zh_*.20in.20Weblate/near/2177452