api: Like zulip-mobile, don't allow connecting to Zulip Server <7.0
This goes rather farther than #1456, which is for bumping this
number just to 5.0. This is OK; we recently bumped zulip-mobile to
use 7.0, in the v27.235 release.

zulip-mobile shows a "nag banner" to nudge server admins to upgrade.
As of v27.235 (PR zulip/zulip-mobile#5922), that banner appears on
7.x servers, the latest 7.x having recently aged out of our 18-month
support window. So we should feel comfortable nudging this number in
zulip-flutter to 8.0 once v27.235 has been out for a little while.

Fixes: #1456