home test: Remove an unnecessary `tester.pump` for a route animation
This test doesn't need to check the widget tree after waiting for
the route animation to complete. The navigator observer is alerted
when the navigation action is dispatched, not when its animation
completes, so it's fine to check pushedRoutes before waiting through
the animation.

(It still needs a Duration.zero wait so that a FakeApiConnection
timer isn't still pending at the end of the test.)