new-dm: Support unselecting a user by tapping chip in input
This isn't specifically mentioned in the Figma, but I think it's
really helpful. Without this, the only way to deselect a user is to
find them again in the list of results. That could be annoying
because the user will often go offscreen in the results list as soon
as you tap it, because we reset the query and scroll state when you
tap it.

Discussion about adding an "x" to the chip, to make this behavior
more visible:
  https://chat.zulip.org/#narrow/channel/516-mobile-dev-help/topic/Follow.20up.20on.20comments.20.20on.20new.20Dm.20Sheet/near/2185229