users: Have userDisplayEmail handle unknown users
Like userDisplayName does.

And remove a null-check `store.getUser(userId)!` at one of the
callers... I think that's *probably* NFC, for the reason given in a
comment ("must exist because UserMentionAutocompleteResult"). But
it's possible this is actually a small bugfix involving a rare race
involving our batch-processing of autocomplete results.

Related: #716