users [nfc]: Use userDisplayName at last non-self-user sites in widgets/
We've now centralized on store.userDisplayName and
store.senderDisplayName for all the code that's responsible for
showing a user's name on the screen, except for a few places we use
`User.fullName` for (a) the self-user and (b) to create an @-mention
for the compose box. The "(unknown user)" and upcoming "Muted user"
placeholders aren't needed for (a) or (b).