msglist: Fetch at first unread, optionally, instead of newest
Fixes #80.

This differs from the behavior of the legacy mobile app, and the
original plan for #80: instead of always using the first unread,
by default we use the first unread only for conversation narrows,
and stick with the newest message for interleaved narrows.

The reason is that once I implemented this behavior (for all narrows)
and started trying it out, I immediately found that it was a lot
slower than fetching the newest message any time I went to the
combined feed, or to some channel feeds -- anywhere that I have a
large number of unreads, it seems.  The request to the server can take
several seconds to complete.

In retrospect it's unsurprising that this might be a naturally slower
request for the database to handle.

In any case, in a view where I have lots of accumulated old unreads,
I don't really want to start from the first unread anyway: I want to
look at the newest history, and perhaps scroll back a bit from there,
to see the messages that are relevant now.

OTOH for someone who makes a practice of keeping their Zulip unreads
clear, the first unread will be relevant now, and they'll likely want
to start from there even in interleaved views.  So make it a setting.

See also chat thread:
  https://chat.zulip.org/#narrow/channel/48-mobile/topic/opening.20thread.20at.20end/near/2192303