unreads test [nfc]: Require isSortedWithoutDuplicates in fillWithMessages
And change some callers slightly, to adapt.

Since fillWithMessages is implemented as handling a sequence of
new-message events, require the messages to be sorted without
duplicates, for realism.

This doesn't change what the model's state looks like for any of the
tests, because messages in the input with the "read" flag are
ignored by model.handleMessageEvent, and we're not changing the
order of the unread messages. Also, the previous commit started
checking the invariant that the model's message-ID lists are sorted.