narrow [nfc]: Make Narrow.containsMessage return nullable bool
Done as preparation to implement a new Narrow subclass for keyword
search.

We can't really implement a client-side "contains-message" predicate
for search; it's up to the server to decide whether a message
matches the search query.

A simple path to representing that, done here, is to just change
Narrow.containsMessage's contract to say that it always returns null
if the narrow is a search narrow.

The app-code changes show just small behavior choices for a
message-list page in a search narrow, which we'll implement soon:
- Message events are ignored
- Outbox messages never appear

Various TODOs in tests; we'll handle all of those as part of adding
the keyword-search Narrow subclass.