msglist [nfc]: Add a notifyListeners, redundant but to simplify reasoning
Seeing that the tests were expecting a notifyListeners call already,
we found that one is already done, in
MessageStore.handleUpdateMessageEvent, so this is redundant.

But because the new call is done synchronously with the existing
call, it doesn't trigger any extra rebuilds, and it helps the reader
by making it clear locally that listeners do indeed get notified.