realm [nfc]: Move processTopicLikeServer to substore, from API code
This way, the method's implementation can look up zulipFeatureLevel
and realmEmptyTopicDisplayName for itself, rather than have them
looked up in advance by the caller and passed in.

Then as a further consequence, the method only looks up
realmEmptyTopicDisplayName if and when it determines it actually needs
that information.  That makes the lookup compatible with the check we
have in the main realmEmptyTopicDisplayName getter to assert that it's
only consulted when it should in fact be needed, on old servers.
A hack in the message store, and a similar hack on the message list,
therefore become unnecessary.