realm [nfc]: Organize realm settings, from current API doc
To identify which fields are realm settings in the API, so that #668
applies to them, I looked at the current API docs:
  https://zulip.com/api/get-events#realm-update_dict

It turns out a couple of items that had been realm settings are now
something else; one that was a server setting (and even lacked "realm"
in the name: maxFileUploadSizeMib) is now a realm setting updated by
events.  So the comments on the implementation fields need updating.
I'll do that in a separate commit, though, so that this one is
more purely moving code around.