emoji [nfc]: Drop setServerEmojiData from main interface
Instead, leave it as a method on EmojiStoreImpl and on the overall
PerAccountStore.

Each of these FooStore types, such as EmojiStore, is implemented by
both a FooStoreImpl class and the overall PerAccountStore.  Nearly all
of these types' methods behave exactly the same on both those types;
the one implementation just forwards to the other.

But this method behaves slightly differently between them: the
PerAccountStore implementation will call notifyListeners, while the
EmojiStoreImpl implementation doesn't (because it can't, not being the
ChangeNotifier itself).

That mismatch, when nearly all the other similar methods have an
exact match, risks confusion.  Mitigate that by removing the method
from the EmojiStore interface, so that the two implementations become
unrelated methods.

(The one other such mismatch is `reconcileMessages`.  We'll deal
with that one too, in a later commit.)