message [nfc]: Drop reconcileMessages from main interface
Much like we did with setServerEmojiData in a recent commit,
and for the same reasons: after setServerEmojiData, this was
the only one of our numerous methods on the various FooStore
interfaces where the two different implementations of the method
actually had different behavior.  That's potentially misleading,
so make the two implementations be unrelated methods instead.