dartdoc: Stop section-divider comments from getting absorbed into docs
Before this change, if you hover in your IDE over an identifier whose
declaration happens to be the first one after one of these section
dividers, the doc shown would begin with a long string of slashes.
After all, the divider line did start with "///".

The choice of "|" as the substitute was fairly arbitrary.  I tried a
space first, and felt it interrupted the divider too much and made it
look like just another line of comments.

Done with a bit of Perl and Git:

  $ perl -i -0pe 's,^\s*//\K(?=//),|,gm' $(git grep -l //// '*.dart')