home [nfc]: Clean up some dead safe-area code from before bottom-tabs
The responsibility for consuming the bottom inset is now external to
all of these *PageBody widgets; it's handled by
Scaffold.bottomNavigationBar. So we don't need to forward a
bottom-inset value down to descendants
(via `SafeArea.bottom: false`), and we don't need to add widgets to
pad the bottom inset; the bottom-inset height in all these places is
zero.