content: Update base KaTeX text style to be explicit
This fixes a bug about `leadingDistribution` where previously it
was taking the default value of
`TextLeadingDistribution.proportional`, now it uses
`TextLeadingDistribution.even` which seems to be the default
strategy used by CSS, and it doesn't look like `katex.scss`
overrides it.

The vertical offsets being updated in tests are because of this fix.

Another potential bug fix is about `textBaseline` where on some
locale systems the default value for this could be
`TextBaseline.ideographic`, and for KaTeX we always want
`TextBaseline.alphabetic`.