action_sheet [nfc]: Rename BottomSheetHeaderPlainText to BottomSheetPlainText
This way, it can be used for purposes other than being a header, such
as in the next commits, when read receipts fail to load, we use this
widget in the middle of bottom sheet to give feedback to the user.

This also adds a TextAlign property for controlling the alignment.