test: Prevent mutation of shared example users
This ensures we don't re-introduce the sort of state leak that
caused #1712.  In three recent commits we fixed every existing
example of this sort of state leak; only three test files had them.

I also looked through all the other top-level fields in this file
`example_data.dart`.  Most are immutable atoms like ints and strings.
There are a handful of remaining points of mutability which could in
principle allow this sort of state leak; so I'll tighten those up too
in the next couple of commits.