action_sheet test: Avoid mutating shared example users
This is the same sort of state leak that caused #1712; this instance
of it just hasn't happened to break any tests for us yet.

We'll soon arrange things so that this sort of state-leaking mutation
causes an immediate error.  This is one of the three total places
where it turns out we had such mutations, including the one we just
fixed in a04b44ede (#1713).

The second of these tests ("no error if recipient was deactivated …")
wasn't actually mutating the shared example user `eg.otherUser`,
because secretly `setupToMessageActionSheet` makes a new User object
with the same user ID and puts that in the store.  Still, it *looked*
like it was; best to do something that clearly looks correct instead.

The first of these tests was indeed mutating `eg.selfUser`, just as it
looks like it's doing.