recent-dms test: Avoid mutating eg.selfUser
Like in the parent commit fixing some action-sheet tests, this is a
state leak that just hasn't happened to break any tests for us yet.

In this case, the fix can also simplify the interface that this
prep helper `setupPage` has with its callers: instead of a specific
feature for setting `fullName` on the self-user, the function can
take an optional User object for the self-user.  Then the individual
test case can do whatever it likes to set up that User object.