test: Prevent mutating shared example ServerEmojiData values
In principle these are subject to the same sort of state leak we've
run into with the shared example User objects, and fixed in recent
commits: these are shared global objects, which don't get discarded or
reset by `testBinding.reset`, and until this commit they were mutable.

The probability that we'd actually end up with such a state leak was
low: ServerEmojiData values never normally get mutated in the app's
data structures (unlike User values), plus there'll probably only ever
be a small number of tests that would have a reason to use these.

But after the preceding couple of commits (notably the one introducing
_ImmutableUser), these represent the last remaining mutable data in
this file's top-level fields.  So let's eliminate that too, and get
to 100% in eliminating the possibility of the #1712 class of bug.