katex [nfc]: Push more parsing into _parseInlineStyles; return Map
This has a small effect on the survey script's list of failure
reasons: in the rare case that the "unexpected shape of inline CSS"
error appears, it now fires before any of the CSS properties in the
same inline style are processed.  That can mean fewer entries added
to unsupportedInlineCssProperties.

This difference is only possible, though, on a KaTeX expression that
is going to reach the same hard failure either way.  So it has no
effect on behavior seen by a user.