katex [nfc]: Consolidate logic for computing overall styles of KatexSpanNode
This just pulls these three pieces of closely-related logic next to
each other.  That will make it easier to refactor them further.

This causes one change in the survey script's list of failure reasons:
when the `delimcenter` class occurs with an inline `top` property, we
now record the unsupported class before reaching the hard fail for the
unsupported property.  This has no user-visible effect, though,
because it can only happen when the expression is going to reach that
hard failure either way.