katex [nfc]: Directly handle expected inline styles on vlist child
There's only a handful of specific properties we expect to see on this
type of span; so handle those explicitly.

In fact, making this list explicit brings to light that there's one
property here which doesn't actually appear on KaTeX's vlist children:
height.  We'll cut that in a separate non-NFC commit.

This will also open up ways to simplify the interaction between this
and the margin-handling logic below.