katex [nfc]: Simplify extracting from attrs map in _parseInlineStyles
This map pattern syntax looks an awful lot like it's saying that
no other keys should be present -- after all, that's what the
corresponding syntax in a list pattern would mean.  So I initially
read this to mean that this code would ignore the inline styles
if any other attribute was present on the element; which wouldn't
be desirable logic.

In fact it's just saying that this one key should be present and
match the given pattern.  But there are simpler ways to say that;
so use one.