katex test [nfc]: Split parsing tests to their own file, like implementation
This way we maintain our usual parallel between the files that tests
live in and the files with the code the tests are about.

There's more we can do to make use of having a specific class for
these.  But we'll do that in separate commits, so that this massive
move commit is as pure of a move as possible.