test: Add find.text analogue with includePlaceholders option
The bulk of this code is copied verbatim from the implementation
of `find.text`, in `package:flutter_test/src/finders.dart`.

The new logic is just the includePlaceholders flag, and passing
it down to where InlineSpan.toPlainText gets called.  The latter
already has a corresponding flag, which does the rest.