login: Reject earlier (at get-server-settings) when server is too old
There's a planned change to make zulipMergeBase required in
GetServerSettingsResult (see #904). When we do that, we'd otherwise
just treat a missing zulipMergeBase on an ancient server the same
way we treat any malformed response. Better to check if the server
is an ancient one we don't support, just like when the /register
response is malformed.