user [nfc]: Split avatar, presence, status widgets out from content.dart
These started here with Avatar, one little widget which we didn't have
a more natural home for and so tucked in with the content widgets.

That gradually grew from 34 lines to 159 lines, across several
widgets.  Then when we built widgets for presence and user status,
we put them here as they're natural companions of the avatar.
Now it's 367 lines in total, none of it really about message content;
so give this all its own separate file.