api: Cut User.isBillingAdmin, obsolete and gone in current servers
We weren't using this; and since it's deprecated, we don't expect
to ever use it in the future.

Because this doesn't appear in the current API docs (except in a
"Changes" note saying how it used to be there), it doesn't need
even a commented-out tombstone, either, like we have for isGuest
and some others.

(This change is NFC on all well-behaved servers; it's not quite NFC
only in that if a server now sends this field but with a value of
an unexpected type, the old code would reject that as malformed and
the new code won't notice or care.)