api [nfc]: Remove a comment about event_types that came from outdated doc
The doc for message_details on this event says the field is present
if the flag is "read", with no mention of event_types. Looking
through zulip/zulip's Git history, it turns out it had been
mentioned in the past, but the mention was removed, in
zulip/zulip@48a1cf04d:

-  Present if `message` and `update_message_flags` are both present in
-  `event_types` and the `flag` is `read` and the `op` is `remove`.
+  Only present if the specified `flag` is `"read"`.

So the comment (and no-op `&& true`) isn't necessary; remove it.

We noticed this because the analyzer started flagging an info-level
`no_literal_bool_comparisons` here.