test [nfc]: Remove the `assert` from serverEmojiDataPopularPlus
The assert was for not letting the coming emoji data collide with
popular emoji data. This change will make it easier for callers to
provide data of any emojis. If the data contains one/more of popular
emojis, the data from serverEmojiDataPopular will replace it, as it is
the source of truth for popular emojis in test code.