NOMERGE autocomplete: Downrank bots even when recent participants
Now that I've spelled this out, complete with its consequences as
seen in the tests, it doesn't actually sound desirable: if there's a
bot participating in the current thread and I start typing its name,
why wouldn't we want that as a top result?

Anyway, this demonstrates how such a change would work out in the
tests, though.