deps: Revert unintentionally-broad upgrade
This reverts commit 55d055f03f0a01f10379aa4e82e6a757b805d75f (#1738).

That change was described as upgrading Flutter and its supporting
libraries, but it looks like it accidentally included other upgrades,
something like `flutter pub upgrade`.  It also didn't include the
CocoaPods updates for those upgrades; see #1746.

So revert that change, to get back to a well-understood state.
Then we can retry the upgrade.