msglist test: Wait through a nav-transition animation in a place we forgot
This should unblock the upstream PR
  https://github.com/flutter/flutter/pull/165832 .
Discussion:
  https://github.com/flutter/flutter/pull/165832#issuecomment-3120915623

(And, for consistency, also use upstream's new
TransitionDurationObserver in a place we were getting a duration
more manually. This weakens the test a bit, by removing checks that
the duration comes from a popped route, with specific details of the
route that was popped. Perhaps we can add a feature to
TransitionDurationObserver to get those checks back, but not a
priority for now.)