deps: Upgrade Flutter to 3.33.0-1.0.pre-1085
And update Flutter's supporting libraries to match.

In particular this gets us TransitionDurationObserver:
  https://github.com/flutter/flutter/pull/171109

We stop just before a certain Dart SDK roll:
  https://github.com/flutter/flutter/commit/fa80cbcbdbdd
because that commit breaks completion in the IDE, which would be
pretty annoying for doing development.  Thanks to Sayed for
spotting that breakage and bisecting it to that commit:
  https://github.com/zulip/zulip-flutter/pull/1755#issuecomment-3124335590
  https://chat.zulip.org/#narrow/channel/516-mobile-dev-help/topic/New.20Flutter.20upgrade.20degrades.20code.20autocompletion/near/2230721