check: Update expected format of Flutter version name
These changed in a recent upstream PR to look a bit more like
the output of `git describe`:
  https://github.com/flutter/flutter/pull/172141

That causes an upgrade by `tools/upgrade flutter-local` past that
upstream change to break an assumption made by this check; the check
notices that up front, and sensibly fails.

So update the check to accept both ways.