katex test [nfc]: Name the KatexExample statics without redundancy
The names beginning "mathBlockKatex…" were appropriate when the
namespace these lived on was the more general ContentExample.
But that prefix doesn't add anything when they're on KatexExample.

(The fact they're in math blocks instead of inline math spans isn't
part of the point of any of these examples.)

Similarly the descriptions can be shortened; and then in some cases
lengthened a bit again to say more about what each example is meant
to exercise or illustrate.