profile: Add a spacer between name/status and email/role
At 4 items high, this is already a bit of a stack of information.
We'll soon add a 5th item, so it'll be good to have it split up.

A logical grouping is to separate the things that are potentially
timely (the name adorned with presence circle and status emoji,
then the status message) from the things that change rarely (the
user's email and role).