emoji_reaction: Remove tooltip on ReactionChip; add semantics label
In a later commit in this series, we'll replace the long-press
action with showing a "view-reactions" sheet. We won't want
something else that's triggered on long-press, like this tooltip,
and the tooltip was sometimes winning the gesture-detection in my
testing at that later commit.

Remove the tooltip, replacing it with a semantics label, which tests
that were relying on the tooltip can use instead.