text: Introduce TextWithLink widget
Fixes part of #1285.

In particular this handles the case we need at the moment for #667,
read receipts: a sentence with a single link in the middle of it.

Other cases noted in #1285 are for a code font; or italics; or
inserting more string fragments before and after a tagged string
(for a Markdown link in quote-and-reply).  There's also #1553
which calls for a sentence with two different links in it (to
use in an empty inbox).