ios: Bump min iOS version to 15.0, as required by latest Firebase iOS SDK
Firebase iOS SDK version 12.0 bumps minimum supported iOS version
to 15.0, see:
  https://firebase.google.com/support/release-notes/ios#version_1200_-_july_15_2025
  https://github.com/firebase/firebase-ios-sdk/issues/14732

So update the min iOS version to 15.0, as we plan to bump our
Firebase dependency soon. Also based on the platform-version
statistics Greg posted:
  https://chat.zulip.org/#narrow/channel/48-mobile/topic/platform.20versions/near/2243725
we could bump it to 16.0, but there isn't specific need for that
right now, so might as well leave that for later.